Item Effects. Active Item Slots. Heal Potion Active Item Slot. Concreate Items Presenter.
Implemented Effects with modular logic.
Added a Heal Potion usable via UI or key input.
Built an ActiveItemsPresenter and a Concrete Items Presenter using MVP and Zenject for input and UI handling.
An EventSystem prefab has been created to generate it once when the GlobalContextScene is loaded.
Move OverUI check from InputListener to InteractionRaycastSystem(propper space for it).